users can now search via query params instead of GUID

+++ Db/creates.sql
@@ -24,7 +24,7 @@ CREATE TABLE demo.products
 (
     product_id CHAR(36) PRIMARY KEY,
     title VARCHAR(32) NOT NULL,
-    price FLOAT NOT NULL DEFAULT 0.00,
+    price DECIMAL(15, 2) NOT NULL DEFAULT 0.00,
     quantity INTEGER UNSIGNED NOT NULL DEFAULT 0,
     created_at DATETIME NOT NULL DEFAULT NOW(),
     updated_at DATETIME NULL
